Handling one or zero items with circular
Having one item with circular set to true resulted in a single item
cloned and scrolling, which is not a desirable condition. Having
zero items with circular set to true resulted in an error. By
enforcing a check more similar to the one performed when circular
is false, we are able to catch this condition.

Additionally, the prev and next buttons are set to disabled in the
case of one or zero items.