Fix for clicking on the next/prev month button hides calendar.
If the next/prev month buttons are hidden via visibility: hidden and the user clicks on the area where the button is after it's been hidden, the click event's target ends up being the root node. The code that checks for this expects the event to be a child of the root node instead of the root node itself, so this patch just adds a check for this.