Update dateinput.js
The test against "msie" UA is weak because most of the browsers can change their UA string so, if a browser is identified as msie but it supports focus method it will fail. 
Maybe in the future versions of IE the focus method would work. In this case we are ready too (however I hope that one of the next IE will support date input natively :-) ).